Search for msbuild on path first
When running the Developer Terminal application, it sets up a MSBuild on
the PATH variable, which has all the needed dependencies to compile the
app as expected.

If we always use the registry msbuild location, the compilation will
keep failing.

Instead of switching completely to use the `msbuild.exe` from PATH, this
commit looks for the executable on the env var first, and fallsback to
the registry if not available.

This should accomodate both envs where the msbuild is setup on the
env var, and the one that relies only on the registry.